requirements.txt
immudb-py==1.1.0
